Split out unix and windows implementations
